성경 장 검색 구현중,,,
